+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ### Changelog
 
+* __1.0.4__:
+    * New:
+        * Support for localized tabbed boxes with `lang-code-tab` instead of name
+    * Fixes:
+        * Undefined `lang-code`s will not overwrite element texts
 * __1.0.3__:
     * New:
         * Support for localization including German and English language